+++ sales-Backend/sales-Backend/src/main/java/com/example/salesBackend/Entity/PG_BENEFIT_MASTER.java
@@ -20,6 +20,6 @@ public class PG_BENEFIT_MASTER {
 
     private String DESCRIPTION;
     private String REMARKS;
-    private Long VIEW_ORDER;
+    private int VIEW_ORDER;
 
 }
