+++ sales-Backend/sales-Backend/src/main/java/com/example/salesBackend/Entity/PG_POLICYLOAN.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -16,12 +17,12 @@ import java.util.Date;
 @Table(name="PG_POLICYLOAN")
 public class PG_POLICYLOAN {
     @Id
-//    @GeneratedValue(strategy=GenerationType.UUID)
+
     @Column(name = "POLICY_NO")
     private String POLICY_NO;
 
     private String LOAN_NO;
     private Date LOAN_DATE;
     private Date LAST_CAPITALIZED_DATE;
-    private Long LOAN_BALANCE;
+    private BigDecimal LOAN_BALANCE;
 }
